fix(sb2040): commission sb2040 changes
heater_bed - needs to be re-enabled
extruder - needs to be added once the thermistor is attached
probe - check the board soldering or remove components for 5v as the
probe does not trigger. The light stays on when moving the tap. It may
require that the CW2 is added to make it dark - i.e. ambient light is
causing issues.